src/test/java/ru/netology/statics/StatsServiceTest.java
package ru.netology.statics;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class StatsServiceTest {

    @Test
    void shouldCalculateSum() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;
        int actual = service.calculateSum(purchases);
        assertEquals(expected, actual);
    }


    @Test
    void shouldCalculateAvergeTotal() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;
        int actual = service.calculateAverageTotal(purchases);
        assertEquals(expected, actual);
    }


    @Test
    void shouldFindMaxMonthSale() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 8;
        int actual = service.findMaxMonthSale(purchases);


        assertEquals(expected, actual);
    }

    @Test
    void shouldFindMinMonthSale() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 9;
        int actual = service.findMinMonthSale(purchases);


        assertEquals(expected, actual);
    }


    @Test
    void shouldCalculateMaxAverageCount() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.calculateMaxAverageCount(purchases);


        assertEquals(expected, actual);
    }

    @Test
    void shouldCalculateMinAverageCount() {
        StatsService service = new StatsService();
        int[] purchases = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.calculateMinAverageCount(purchases);


        assertEquals(expected, actual);
    }
}











